Update as the requirements in email
- Search history: it’s doesn’t update the number of passengers on
search view when the dates are too old. Use case: my search screen is
set with 3 passengers. I then go to the search history and tap on a
search with different amount of passengers. If the date is too old, the
app sends me to the search view but the number of passenger is still
set to 3, instead of the relevant number.

- Flight length bar: Circles are sometime overlapping the airport code
(see print screen), or overlapping each others.

- Filters: With Bangkok, there is 2 times Survanabhumi, shouldn’t the
first one be “Bangkok All airports”? Also, the app let me deselect the
airports but then there is no result and the screen is empty. Should we
force to at least one airport selected?

- Cabin Class: We forgot to include it in the design, since the first
API we used didn’t have it