Fix minor buges - Feedback Jan, 24th
1. add space between number and “room”
———> DONE
2. we should round up the price to avoid digits
———> DONE

3. Can we/should we have the markers in the middle of the view?
———> DONE
4. What is this “Search Here” button for? When I move the map and tap
the button, I don’t see new hotels showing up, even downtown Bangkok
(note: I had a crash when I closed the map but cannot reproduce, will
keep testing that)
———> Search This Area Button is for searching hotels around that area.
If you don’t see any hotel there, simply there are just no hotels.
Currently, our API doesn’t provide this feature. However, I used the
latitude and longitude to show the hotels on MAP, to let people easily
find the hotel near where they want.
———> I also noticed the crash, it is due to memory overload, it is
because some city they have a lot of markers in the same area, I added
it all to the map, leaded to memory overloading, then the app crashed
itself. I did solve that by using the algorithm that only show maximum
of first 50 hotels in the return list, and only show the hotels in the
radius of 5 kms from center of the map view. After that, the map will
be moving itself to the center hotel of the hotels.

5. Add margin below the button (should be in the design)
———> DONE
6. should we switch tab to hotel or flight when we launch a search from
history?
———> Basically, I don’t think we should switch. The reason why I
implemented that way is that user can search directly from History
Searches, and back to the searches directly if they don’t wanna go
deeper. If we jump in the disturbance of tabs, it will be the mess for
users.

7. Images are sometime slow to load, can we have a loading animation or
at least a light gray placeholder so the layout looks ok from the
beginning?
———> DONE, I use Image place holder for this case.